// shared/common/enums/caching.enum.ts
export enum CacheTypes {
  LOCALIZATION = 'LOCALIZATION',
  CONTENTURI = 'CONTENTURI',
  TOKENURI = 'TOKENURI',
  TOKEN = 'TOKEN',
  PAGE = 'PAGE',
  PAGEMETA = 'PAGEMETA',
  COMPONENT = 'COMPONENT',
  COMPONENTMETA = 'COMPONENTMETA',
  COMPONENTLIST = 'COMPONENTLIST',
  BINARY = 'BINARY',
  BINARYTYPE = 'BINARYTYPE',
  BINARYLASTPUBLISHDATE = 'BINARYLASTPUBLISHDATE',
  LABEL = 'LABEL',
  CMSCONFIGURATION = 'CMSCONFIGURATION',
  CATEGORY = 'CATEGORY',
  CATEGORYWITHKEYWORD = 'CATEGORYWITHKEYWORD',
  KEYWORD = 'KEYWORD',
  HEADER = 'HEADER',
  FOOTER = 'FOOTER',
  NAVIGATION = 'NAVIGATION',
  REDIRECTS = 'REDIRECTS',
  ALL = 'ALL',
}